requirements.txt
black==24.10.0
numpy
tabulate==0.9.0
pyright==1.1.387
marshmallow-dataclass